requirements.txt
requests
furl
ansicolors
beautifulsoup4
pyotp
